create helper for provider requests in tests
