+++ extension/test-utils/wallet/index.ts
@@ -1,3 +1,4 @@
 export { connectMockWallet } from './connectMockWallet'
 export { MockProvider } from './MockProvider'
+export { mockProviderRequest } from './mockProviderRequest'
 export { mockRpcRequest } from './mockRpcRequest'
